+++ lib/http_wrapper/http_wrapper.rb
@@ -1,9 +1,9 @@
 require 'net/https'
 
 class HTTPWrapper
-  KNOWN_OPTIONS_KEYS = [:timeout, :ca_file, :validate_ssl_cert].freeze
+  KNOWN_OPTIONS_KEYS = [:timeout, :verify_cert, :logger, :max_redirects].freeze
 
-  attr_accessor :timeout, :ca_file, :validate_ssl_cert
+  attr_accessor :timeout, :verify_cert, :logger, :max_redirects
 
   def initialize(options = {})
     unknown_options = options.keys - KNOWN_OPTIONS_KEYS
@@ -12,9 +12,10 @@ class HTTPWrapper
       raise UnknownParameterError.new "Unknown options: #{unknown_options.join(', ')}"
     end
 
-    @timeout           = options.fetch(:timeout) { 10 }
-    @ca_file           = options.fetch(:ca_file) { nil }
-    @validate_ssl_cert = options.fetch(:validate_ssl_cert) { false }
+    @timeout       = options.fetch(:timeout) { 10 }
+    @verify_cert   = options.fetch(:verify_cert) { true }
+    @logger        = options.fetch(:logger) { nil }
+    @max_redirects = options.fetch(:max_redirects) { 10 }
   end
 
   [:get, :post, :put, :delete].each do |method|
@@ -38,13 +39,15 @@ class HTTPWrapper
     response.response['set-cookie']
   end
 
-  def get_response(request, redirects_limit = 10)
+  private
+
+  def get_response(request, redirects_limit = @max_redirects)
     raise TooManyRedirectsError.new 'Too many redirects!' if redirects_limit == 0
 
     response = perform_request request
 
     if response.kind_of? Net::HTTPRedirection
-      request.url = response['location']
+      request.uri = response['location']
       response = get_response request, redirects_limit - 1
     end
 
@@ -52,26 +55,21 @@ class HTTPWrapper
   end
 
   def perform_request(request)
-    connection = create_connection request.url
-    connection.request request.create
+    connection = create_connection_for request.uri
+    request.perform_using connection
   end
 
-  def create_connection(uri)
+  def create_connection_for(uri)
     connection = Net::HTTP.new uri.host, uri.port
-    connection.read_timeout = timeout
-    connection.open_timeout = timeout
+    connection.read_timeout = @timeout
+    connection.open_timeout = @timeout
 
-    if uri.scheme == 'https'
+    if uri.kind_of? URI::HTTPS
       connection.use_ssl = true
-
-      if validate_ssl_cert
-        connection.ca_file = ca_file
-        connection.verify_mode = OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
-      else
-        connection.verify_mode = OpenSSL::SSL::VERIFY_NONE
-      end
+      connection.verify_mode = OpenSSL::SSL::VERIFY_NONE unless @verify_cert
     end
 
+    connection.set_debug_output(@logger)
     connection
   end
 end